strtrim, makefile

+++ ft_strtrim.c
@@ -6,11 +6,12 @@
 /*   By: jyou <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/15 16:05:56 by jyou              #+#    #+#             */
-/*   Updated: 2020/11/01 17:34:30 by jyou             ###   ########.fr       */
+/*   Updated: 2020/11/10 17:07:09 by jyou             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdio.h>
 
 static size_t		find_begin(char const *s1, char const *set)
 {
@@ -61,14 +62,14 @@ char				*ft_strtrim(char const *s1, char const *set)
 	char		*str;
 	size_t		begin;
 	size_t		end;
-	int			str_len;
+	size_t		str_len;
 
 	if (!s1)
 		return (NULL);
 	if (!set)
 		return (ft_strdup(s1));
 	begin = find_begin(s1, set);
-	if (begin == ft_strlen(s1))
+	if (begin == (size_t)ft_strlen(s1))
 		return (ft_strdup(""));
 	end = find_end(s1, set);
 	str_len = end - begin + 2;
